Move test_ephys_trials tests to test_ephys_fpga; remove old functions
